+++ R/reexports.R
@@ -6,6 +6,10 @@ ape::rtree
 ##' @export
 ape::read.tree
 
+##' @importFrom ape read.nexus
+##' @export
+ape::read.nexus
+
 ##' @importFrom tidytree groupOTU
 ##' @export
 tidytree::groupOTU
@@ -13,6 +17,18 @@ tidytree::groupOTU
 ##' @importFrom tidytree groupClade
 ##' @export
 tidytree::groupClade
+
+##' @importFrom dplyr collapse
+##' @export
+dplyr::collapse
+
+##' @importFrom ggplot2 fortify
+##' @export
+ggplot2::fortify
+
+##' @importFrom ggplot2 ggplot
+##' @export
+ggplot2::ggplot
 
 ##' @importFrom ggplot2 xlim
 ##' @export
@@ -43,6 +59,59 @@ ggplot2::geom_label
 ggplot2::geom_point
 
 
+##' @importFrom treeio read.beast
+##' @export
+treeio::read.beast
 
+##' @importFrom treeio read.codeml
+##' @export
+treeio::read.codeml
 
+##' @importFrom treeio read.codeml_mlc
+##' @export
+treeio::read.codeml_mlc
+
+##' @importFrom treeio read.hyphy
+##' @export
+treeio::read.hyphy
+
+##' @importFrom treeio read.jplace
+##' @export
+treeio::read.jplace
+
+##' @importFrom treeio read.jtree
+##' @export
+treeio::read.jtree
+
+##' @importFrom treeio read.mrbayes
+##' @export
+treeio::read.mrbayes
+
+##' @importFrom treeio read.newick
+##' @export
+treeio::read.newick
+
+##' @importFrom treeio read.nhx
+##' @export
+treeio::read.nhx
+
+##' @importFrom treeio read.paml_rst
+##' @export
+treeio::read.paml_rst
+
+##' @importFrom treeio read.phylip
+##' @export
+treeio::read.phylip
+
+##' @importFrom treeio read.phyloT
+##' @export
+treeio::read.phyloT
+
+##' @importFrom treeio read.r8s
+##' @export
+treeio::read.r8s
+
+##' @importFrom treeio read.raxml
+##' @export
+treeio::read.raxml