+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggtree 1.11.4
 
++ re-implement collapse as collapse.ggtree method by importing dplyr::collapse
+  generic to prevent function name collision (2018-01-03, Wed)
 + update treeVisualization vignette, with more layout examples added (2017-12-22, Fri)
 + update vignette (2017-12-21, Thu)
     - remove ggtreeUtilities.Rmd